refactor: clean up LimitOrders page
