+++ apps/cowswap-frontend/src/modules/limitOrders/index.ts
@@ -1,11 +1,12 @@
-export * from './containers/LimitOrdersWidget'
+export * from './const/trade'
 export * from './containers/ChartWidget'
-export * from './updaters/QuoteObserverUpdater'
-export * from './updaters/InitialPriceUpdater'
-export * from './updaters/ExecutionPriceUpdater'
-export * from './updaters/FillLimitOrdersDerivedStateUpdater'
-export * from './updaters/SetupLimitOrderAmountsFromUrlUpdater'
+export * from './containers/LimitOrdersWidget'
 export * from './state/limitOrdersRawStateAtom'
 export * from './state/limitOrdersSettingsAtom'
+export * from './updaters/AlternativeLimitOrderUpdater'
+export * from './updaters/ExecutionPriceUpdater'
+export * from './updaters/FillLimitOrdersDerivedStateUpdater'
+export * from './updaters/InitialPriceUpdater'
+export * from './updaters/QuoteObserverUpdater'
+export * from './updaters/SetupLimitOrderAmountsFromUrlUpdater'
 export { useIsWidgetUnlocked } from './hooks/useIsWidgetUnlocked'
-export * from './const/trade'
