--- apps/cowswap-frontend/src/pages/AlternativeOrder/index.tsx
@@ -1,1 +1,0 @@
-export { AlternativeLimitOrder } from './AlternativeLimitOrder'
