+++ apps/cowswap-frontend/src/pages/LimitOrders/index.tsx
@@ -1,59 +1,32 @@
-import { useWalletInfo } from '@cowprotocol/wallet'
-
-import { useOrders } from 'legacy/state/orders/hooks'
-
 import { AppDataUpdater } from 'modules/appData'
 import {
+  AlternativeLimitOrderUpdater,
   ExecutionPriceUpdater,
   FillLimitOrdersDerivedStateUpdater,
   InitialPriceUpdater,
   LIMIT_ORDER_SLIPPAGE,
-  LimitOrdersWidget,
   QuoteObserverUpdater,
   SetupLimitOrderAmountsFromUrlUpdater,
-  useIsWidgetUnlocked,
 } from 'modules/limitOrders'
-import { OrdersTableWidget } from 'modules/ordersTable'
-import { TabOrderTypes } from 'modules/ordersTable/pure/OrdersTableContainer'
-import * as styledEl from 'modules/trade/pure/TradePageLayout'
 
 import { useIsAlternativeOrderModalVisible } from 'common/state/alternativeOrder'
-import { AlternativeLimitOrder } from 'pages/AlternativeOrder'
-import { UiOrderType } from 'utils/orderUtils/getUiOrderType'
 
-function LimitOrderPage() {
-  const { chainId, account } = useWalletInfo()
-  const allLimitOrders = useOrders(chainId, account, UiOrderType.LIMIT)
+import { AlternativeLimitOrder } from './AlternativeLimitOrder'
+import { RegularLimitOrders } from './RegularLimitOrders'
 
-  const isUnlocked = useIsWidgetUnlocked()
+export default function LimitOrderPage() {
+  const isAlternative = useIsAlternativeOrderModalVisible()
 
   return (
     <>
       <AppDataUpdater orderClass="limit" slippage={LIMIT_ORDER_SLIPPAGE} />
       <QuoteObserverUpdater />
+      {isAlternative && <AlternativeLimitOrderUpdater />}
       <FillLimitOrdersDerivedStateUpdater />
-      <SetupLimitOrderAmountsFromUrlUpdater />
+      {!isAlternative && <SetupLimitOrderAmountsFromUrlUpdater />}
       <InitialPriceUpdater />
       <ExecutionPriceUpdater />
-      <styledEl.PageWrapper isUnlocked={isUnlocked}>
-        <styledEl.PrimaryWrapper>
-          <LimitOrdersWidget />
-        </styledEl.PrimaryWrapper>
-
-        <styledEl.SecondaryWrapper>
-          <OrdersTableWidget
-            displayOrdersOnlyForSafeApp={false}
-            orderType={TabOrderTypes.LIMIT}
-            orders={allLimitOrders}
-          />
-        </styledEl.SecondaryWrapper>
-      </styledEl.PageWrapper>
+      {isAlternative ? <AlternativeLimitOrder /> : <RegularLimitOrders />}
     </>
   )
-}
-
-export default function Page() {
-  const isAlternative = useIsAlternativeOrderModalVisible()
-
-  return isAlternative ? <AlternativeLimitOrder /> : <LimitOrderPage />
 }
